+++ ExamPart.h
@@ -3,18 +3,18 @@
 class ExamPart {
 
 private:
-    double ratio = 0;
+    int ratio = 0;
     int completeness = 0;
 
 public:
-    ExamPart(double _ratio):ratio(_ratio){};
+    ExamPart(int _ratio):ratio(_ratio){};
     ExamPart(){};
     ~ExamPart();
 
-    double getPartRatio();
+    int getPartRatio();
     int getPartCompleteness();
 
-    void setPartRatio(double);
+    void setPartRatio(int);
     void setPartCompleteness(int);
 
     void addPartCompleteness(int);
